Moved all clients to their own folder + split clients into partials.
