+++ DSharpPlus/Clients/BaseDiscordClient.cs
